python_codes/pyproject.toml
[project]
name = "python-codes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyserial>=3.5",
]
